Tracing Wholesale Crude Petroleum Prices through an ARIMA Model
Abstract

This project attempts to trace the trend in the wholesale crude oil prices over 48 months, between January, 2013 and December, 2016. The analysis is based on data from the Wholesale Price Index, published monthly by the Office of the Economic Adviser, under the Ministry of Commerce and Industry of India. Critically analysing this trend is necessary, as crude petroleum prices are often indicative of the overall health of the economy.